64 bit update.

+++ VBitmap.h
@@ -46,7 +46,7 @@ typedef unsigned long	Pixel32;
 typedef unsigned char	Pixel8;
 typedef long			PixCoord;
 typedef	long			PixDim;
-typedef	long			PixOffset;
+typedef	long long		PixOffset; // was just typdef long
 
 #ifdef VDEXT_VIDEO_FILTER
 #define NOVTABLE __declspec(novtable)
